Nutriet/requirements.txt
# ══════════════════════════════════════════════════════════════
#  NutriET — Dependencias Python
# ══════════════════════════════════════════════════════════════

# Django Core
Django==5.2.8
asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.2
typing_extensions==4.15.0

# Authentication & Authorization
django-allauth==65.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2

# Forms & Validation
django-formtools==2.5.1

# API & HTTP
requests==2.32.5
httpx==0.25.2

# Translation
deep-translator==1.11.4

# Image Processing
pillow==12.0.0

# Web Scraping
beautifulsoup4==4.14.3
soupsieve==2.8

# Firebase
firebase-admin==6.4.0

# Gemini AI
google-generativeai==0.8.5

# Data Processing & ML
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.2

# Task Queue & Async (APScheduler sin Celery — más fácil de desplegar)
apscheduler==3.10.4

# Celery + Redis (opcional, comentar si no usas Redis)
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Environment Variables
python-dotenv==1.0.1
python-decouple==3.8

# Database Utilities
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# CORS Support
django-cors-headers==4.3.1

# Serialization
djangorestframework-camel-case==1.3.3

# Date/Time utilities
python-dateutil==2.8.2

# Security
cryptography==42.0.5

# HTTP & Security
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
